Move northstar-specific conformance tests back to northstar
