+++ change/@fluentui-react-conformance-a2e324fa-5aa6-4d14-91cd-c78679da2184.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "BREAKING CHANGE: Remove some tests and options specific to `@fluentui/react-northstar`",
+  "packageName": "@fluentui/react-conformance",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "minor"
+}
