+++ packages/react-avatar/src/common/isConformant.ts
@@ -6,9 +6,7 @@ export function isConformant<TProps = {}>(
   testInfo: Omit<IsConformantOptions<TProps>, 'componentPath'> & { componentPath?: string },
 ) {
   const defaultOptions: Partial<IsConformantOptions<TProps>> = {
-    asPropHandlesRef: true,
     componentPath: module!.parent!.filename.replace('.test', ''),
-    disabledTests: ['has-docblock'],
     extraTests: griffelTests as TestObject<TProps>,
   };
