+++ packages/react-button/src/common/isConformant.ts
@@ -6,11 +6,9 @@ export function isConformant<TProps = {}>(
   testInfo: Omit<IsConformantOptions<TProps>, 'componentPath'> & { componentPath?: string },
 ) {
   const defaultOptions: Partial<IsConformantOptions<TProps>> = {
-    asPropHandlesRef: true,
     componentPath: module!.parent!.filename.replace('.test', ''),
-    disabledTests: ['has-docblock', 'kebab-aria-attributes'],
+    disabledTests: ['kebab-aria-attributes'],
     extraTests: griffelTests as TestObject<TProps>,
-    skipAsPropTests: true,
   };
 
   baseIsConformant(defaultOptions, testInfo);
