+++ packages/react-conformance/src/index.ts
@@ -1,2 +1,3 @@
 export { isConformant } from './isConformant';
 export * from './types';
+export { getComponent } from './utils';
