+++ packages/react-conformance/src/utils/index.ts
@@ -1,3 +1,5 @@
+export * from './errorMessages';
+export * from './getCallbackArguments';
 export * from './getComponent';
-export { default as consoleUtil } from './consoleUtil';
-export { default as parseDocblock } from './parseDocblock';
+export { getPackagePath } from './getPackagePath';
+export * from './validateCallbackArguments';
