+++ packages/react-focus/src/components/FocusZone/FocusZone.test.tsx
@@ -84,10 +84,9 @@ describe('FocusZone', () => {
     Component: FocusZone,
     displayName: 'FocusZone',
     disabledTests: [
-      // Their is no existing top level FocusZone.ts file.
+      // There is no top level FocusZone.ts file.
       'has-top-level-file',
     ],
-    asPropHandlesRef: true,
     elementRefName: 'elementRef',
     testOptions: {
       'consistent-callback-names': { ignoreProps: ['onActiveElementChanged'] },
