+++ packages/react-link/src/common/isConformant.ts
@@ -6,12 +6,9 @@ export function isConformant<TProps = {}>(
   testInfo: Omit<IsConformantOptions<TProps>, 'componentPath'> & { componentPath?: string },
 ) {
   const defaultOptions: Partial<IsConformantOptions<TProps>> = {
-    asPropHandlesRef: true,
     componentPath: module!.parent!.filename.replace('.test', ''),
-    disabledTests: [`has-docblock`],
     extraTests: griffelTests as TestObject<TProps>,
     // TODO: To fix as part of https://github.com/microsoft/fluentui/issues/19522
-    skipAsPropTests: true,
   };
 
   baseIsConformant(defaultOptions, testInfo);
