+++ packages/react-provider/src/components/FluentProvider/FluentProvider.test.tsx
@@ -18,7 +18,6 @@ describe('FluentProvider', () => {
     disabledTests: ['component-handles-classname'],
     Component: FluentProvider,
     displayName: 'FluentProvider',
-    skipAsPropTests: true,
     helperComponents: [ProviderContext.Provider, TextDirectionProvider],
   });
