+++ packages/react-tooltip/src/components/Tooltip/Tooltip.test.tsx
@@ -33,9 +33,6 @@ describe('Tooltip', () => {
       'component-has-root-ref',
       'component-handles-classname',
       'component-has-static-classname',
-      'as-renders-fc',
-      'as-passes-as-value',
-      'as-renders-html',
     ],
   });
