+++ packages/react/src/components/Announced/Announced.test.tsx
@@ -40,7 +40,6 @@ describe('Announced', () => {
     Component: Announced,
     displayName: 'Announced',
     componentPath: path.join(__dirname, 'Announced.ts'),
-    skipAsPropTests: false,
     // Problem: Ref isn't passed.
     // Solution: Ref should be added and passed onto the root.
     disabledTests: ['component-handles-ref', 'component-has-root-ref'],
