+++ packages/react/src/components/Breadcrumb/Breadcrumb.test.tsx
@@ -99,15 +99,9 @@ describe('Breadcrumb', () => {
   isConformant({
     Component: Breadcrumb,
     displayName: 'Breadcrumb',
-    skipAsPropTests: false,
     // Problem: Doesn’t handle ref.
     // Solution: Add a ref to the root element.
-    disabledTests: [
-      'component-handles-ref',
-      'component-has-root-ref',
-      'component-handles-classname',
-      'as-passes-as-value',
-    ],
+    disabledTests: ['component-handles-ref', 'component-has-root-ref', 'component-handles-classname'],
   });
 
   it('renders items with expected element type', () => {
