+++ packages/react/src/components/Label/Label.test.tsx
@@ -19,7 +19,6 @@ describe('Label', () => {
   isConformant({
     Component: Label,
     displayName: 'Label',
-    skipAsPropTests: false,
     // Problem: Ref is not supported
     // Solution: Convert to FunctionComponent and support using forwardRef
     disabledTests: ['component-handles-ref', 'component-has-root-ref'],
