+++ packages/react/src/components/Toggle/Toggle.test.tsx
@@ -49,7 +49,6 @@ describe('Toggle', () => {
     Component: Toggle,
     displayName: 'Toggle',
     targetComponent: Toggle,
-    skipAsPropTests: false,
   });
 
   it('renders aria-label', () => {
